Move CheckFinishedStatus to test-utils
